refactoring: extract todo components and add open api for them
