+++ components/todo/constants/filter-options.ts
@@ -1,0 +1,16 @@
+import { UIButtonsGroupOptionModel } from "@/ui/types";
+
+export const filterOptions: UIButtonsGroupOptionModel[] = [
+  {
+    label: "all",
+    value: "all",
+  },
+  {
+    label: "completed",
+    value: "completed",
+  },
+  {
+    label: "uncompleted",
+    value: "uncompleted",
+  },
+];
