+++ components/todo/index.ts
@@ -1,0 +1,1 @@
+export { Todo } from "./todo";
